+++ src/main/java/com/assetflow/loan/Loan.java
@@ -18,7 +18,6 @@ public class Loan {
     @Column(name = "loan_id")
     private Long id;
 
-
     @Enumerated(EnumType.STRING)
     private LoanStatus loanStatus;