+++ src/main/java/com/assetflow/loan/dto/MyLoanListResponse.java
@@ -1,8 +1,6 @@
 package com.assetflow.loan.dto;
 
-import com.assetflow.asset.AssetItem;
 import com.assetflow.loan.LoanStatus;
-import com.assetflow.member.Member;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
@@ -10,10 +8,9 @@ import java.time.LocalDate;
 
 @Getter
 @AllArgsConstructor
-public class LoanCreateResponse {
+public class MyLoanListResponse {
     private Long loanId;
     private LoanStatus loanStatus;
-    private Long memberId;
     private Long assetItemId;
     private LocalDate loanDate;
     private LocalDate dueDate;